feat(flux2): add Klein/Qwen3 variant support and encoder filtering
Backend:
- Add klein_4b/klein_9b variants for FLUX.2 Klein models
- Add qwen3_4b/qwen3_8b variants for Qwen3 encoder models
- Validate encoder variant matches Klein model (4B↔4B, 9B↔8B)
- Auto-detect Qwen3 variant from hidden_size during probing

Frontend:
- Show variant field for all model types in ModelView
- Filter Qwen3 encoder dropdown to only show compatible variants
- Update variant type definitions (zFlux2VariantType, zQwen3VariantType)
- Remove unused exports (isFluxDevMainModelConfig, isFlux2Klein9BMainModelConfig)

+++ invokeai/backend/model_manager/load/model_loaders/z_image.py
@@ -595,7 +595,10 @@ class Qwen3EncoderCheckpointLoader(ModelLoader):
                 # Dequantize: convert to float and multiply by scale
                 # Handle block-wise quantization (e.g., FP4 with block_size=8)
                 # where scale has shape [weight_dim / block_size, ...]
-                weight_float = weight.to(torch.float32)
+                # Note: Float8 types (e.g., float8_e4m3fn) require .float() instead of .to(torch.float32)
+                # as PyTorch doesn't support direct type promotion for Float8 types
+                weight_float = weight.float()
+                scale = scale.float()
                 if scale.shape != weight_float.shape and scale.numel() > 1:
                     # Block-wise quantization: need to expand scale to match weight shape
                     # Find which dimension differs and repeat scale along that dimension
